tests/wpt/web-platform-tests/domparsing/innerhtml-04.html

<!DOCTYPE html>
<title>innerHTML in HTML</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="https://w3c.github.io/DOM-Parsing/#widl-Element-innerHTML">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
function testIsChild(p, c) {
  assert_equals(p.firstChild, c);
  assert_equals(c.parentNode, p);
}
test(function() {
  var p = document.createElement('p');
  var b = p.appendChild(document.createElement('b'));
  var t = b.appendChild(document.createTextNode("foo"));
  testIsChild(p, b);
  testIsChild(b, t);
  assert_equals(t.data, "foo");
  p.innerHTML = "";
  testIsChild(b, t);
  assert_equals(t.data, "foo");
}, "innerHTML should leave the removed children alone.")
</script>
